feat: creating the routes
