+++ src/db/db.json
@@ -1,1 +1,1 @@
-{"contacts":[{"id":"8a5a4b2f-9054-4d03-ad9f-64c6a3ae5b42","name":"marcio","mobilePhone":"1234","email":"marcio#example.com"}]}
+{"contacts":[{"id":"8a5a4b2f-9054-4d03-ad9f-64c6a3ae5b42","name":"marcio","mobilePhone":"1234","email":"marcio#example.com"},{"id":"3ed01e4e-f415-4f0b-9a97-fd0f2bafac19","name":"Lavinia","mobilePhone":"1234","email":"lavinia@example.com"},{"id":"c2bc4448-3f33-48d2-962f-62b9d7f9d8c3","name":"Lavinia","mobilePhone":"1234","email":"lavinia@example.com"}]}
