+++ src/utils/consts.js
@@ -1,0 +1,2 @@
+export const CONTANCT_URL = "/contact";
+export const TABLE_NAME = "contacts";
